Move VCR conftest to tests.vcr
