+++ tests/challenges/conftest.py
@@ -6,8 +6,7 @@ from _pytest.config.argparsing import Parser
 from _pytest.fixtures import FixtureRequest
 
 from tests.challenges.challenge_decorator.challenge import Challenge
-from tests.integration.conftest import BASE_VCR_CONFIG
-from tests.vcr.vcr_filter import before_record_response
+from tests.vcr import BASE_VCR_CONFIG, before_record_response
 
 
 def before_record_response_filter_errors(
