+++ tests/challenges/debug_code/test_debug_code_challenge_a.py
@@ -7,9 +7,7 @@ from autogpt.agent import Agent
 from autogpt.commands.execute_code import execute_python_file
 from autogpt.commands.file_operations import append_to_file, write_to_file
 from autogpt.config import Config
-from tests.challenges.challenge_decorator.challenge_decorator import (
-    challenge,
-)
+from tests.challenges.challenge_decorator.challenge_decorator import challenge
 from tests.challenges.utils import run_interaction_loop
 from tests.utils import requires_api_key
