+++ tests/challenges/information_retrieval/test_information_retrieval_challenge_a.py
@@ -3,9 +3,7 @@ from pytest_mock import MockerFixture
 
 from autogpt.commands.file_operations import read_file
 from autogpt.config import Config
-from tests.challenges.challenge_decorator.challenge_decorator import (
-    challenge,
-)
+from tests.challenges.challenge_decorator.challenge_decorator import challenge
 from tests.challenges.utils import run_interaction_loop
 from tests.utils import requires_api_key
