+++ tests/challenges/memory/test_memory_challenge_b.py
@@ -4,9 +4,7 @@ from pytest_mock import MockerFixture
 from autogpt.agent import Agent
 from autogpt.commands.file_operations import read_file, write_to_file
 from autogpt.config import Config
-from tests.challenges.challenge_decorator.challenge_decorator import (
-    challenge,
-)
+from tests.challenges.challenge_decorator.challenge_decorator import challenge
 from tests.challenges.utils import generate_noise, run_interaction_loop
 from tests.utils import requires_api_key
